requirements.txt
torch
pillow
gradio
torchvision
git+https://github.com/openai/CLIP.git